% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{coef.signpen}
\alias{coef.signpen}
\title{get coefficients or make coefficient predictions from an "signpen" object.}
\usage{
\method{coef}{signpen}(object, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{signpen}} model object.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which predictions
are required. Default is the entire sequence used to create the model.}

\item{\dots}{not used. Other arguments to predict.}
}
\value{
The coefficients at the requested values for \code{lambda}.
}
\description{
Computes the coefficients at the requested values for \code{lambda} from a
  fitted \code{\link{signpen}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s}
  is not in the lambda sequence used for fitting the model, the \code{coef}
  function will use linear interpolation to make predictions. The new values
  are interpolated using a fraction of coefficients from both left and right
  \code{lambda} indices.
}
\keyword{consistency}
\keyword{penalty}
\keyword{perceptron}
\keyword{sign}
