% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.signpen.R
\name{cv.signpen}
\alias{cv.signpen}
\alias{cv.signlog}
\alias{cv.signls}
\title{Cross-validation for signpen}
\usage{
cv.signpen(x, y, lambda = NULL, pred.loss = c("misclass", "loss"),
  nfolds = 5, foldid, ...)
}
\arguments{
\item{x}{matrix of predictors, of dimension \eqn{n \times p}{n*p}; each row
is an observation vector.}

\item{y}{response variable. This argument should be quantitative for least squares, 
and a two-level factor for logistic regression.}

\item{lambda}{optional user-supplied lambda sequence; default is
\code{NULL}, and \code{\link{signpen}} chooses its own sequence.}

\item{pred.loss}{loss to use for cross-validation error. Valid options are:
\itemize{ \item \code{"loss"} for classification, margin based loss
function.  \item \code{"misclass"} for classification, it gives
misclassification error.  \item \code{"L1"} for regression, mean square
error used by least squares regression \code{loss="ls"}, it measure the
deviation from the fitted mean to the response.  \item \code{"L2"} for
regression, mean absolute error used by least squares regression
\code{loss="ls"}, it measure the deviation from the fitted mean to the
response.  } Default is \code{"loss"}.}

\item{nfolds}{number of folds; default is 5. 
Smallest value allowable is \code{nfolds=3}.}

\item{foldid}{an optional vector of values between 1 and \code{nfold}
identifying what fold each observation is in. 
If supplied, \code{nfold} can be missing.}

\item{\dots}{other arguments that can be passed to signpen.}
}
\value{
an object of class \code{\link{cv.signpen}} is returned, which is a
  list with the ingredients of the cross-validation fit.  \item{lambda}{the
  values of \code{lambda} used in the fits.} \item{cvm}{the mean
  cross-validated error - a vector of length \code{length(lambda)}.}
  \item{cvsd}{estimate of standard error of \code{cvm}.} \item{cvupper}{upper
  curve = \code{cvm+cvsd}.} \item{cvlower}{lower curve = \code{cvm-cvsd}.}
  \item{name}{a text string indicating type of measure (for plotting
  purposes).} \item{signpen.fit}{a fitted \code{\link{signpen}} object for the
  full data.} \item{lambda.min}{The optimal value of \code{lambda} that gives
  minimum cross validation error \code{cvm}.} \item{lambda.1se}{The largest
  value of \code{lambda} such that error is within 1 standard error of the
  minimum.}
}
\description{
Does k-fold cross-validation for signpen and returns a
value for \code{lambda}. This function is modified based on the \code{cv}
function from the \code{glmnet} package.
}
\details{
The function runs \code{\link{signpen}} \code{nfolds}+1 times; the first to
get the \code{lambda} sequence, and then the remainder to compute the fit
with each of the folds omitted. The average error and standard deviation
over the folds are computed.
}
\keyword{consistency}
\keyword{penalty}
\keyword{perceptron}
\keyword{sign}
