% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.cv.signpen}
\alias{predict.cv.signpen}
\title{make predictions from a "cv.signpen" object.}
\usage{
\method{predict}{cv.signpen}(object, newx, s = c("lambda.1se",
  "lambda.min"), ...)
}
\arguments{
\item{object}{fitted \code{\link{cv.signpen}} object.}

\item{newx}{matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix. See documentation for \code{predict.signpen}.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the value \code{s="lambda.1se"} stored
on the CV object. Alternatively \code{s="lambda.min"} can be used. If
\code{s} is numeric, it is taken as the value(s) of \code{lambda} to be
used.}

\item{\dots}{not used. Other arguments to predict.}
}
\value{
The returned object depends on the \dots{} argument which is passed
  on to the \code{\link{predict}} method for \code{\link{signpen}} objects.
}
\description{
This function makes predictions from a cross-validated \code{signpen} model,
  using the stored \code{"signpen.fit"} object, and the optimal value chosen
  for \code{lambda}.
}
\details{
This function makes it easier to use the results of cross-validation to make
  a prediction.
}
\keyword{consistency}
\keyword{penalty}
\keyword{perceptron}
\keyword{sign}
