% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.signpen}
\alias{print.signpen}
\title{print a signpen object}
\usage{
\method{print}{signpen}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{\link{signpen}} object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
a two-column matrix, the first columns is the number of nonzero
  group counts and the second column is \code{Lambda}.
}
\description{
Print the nonzero group counts at each lambda along the signpen path.
}
\details{
Print the information about the nonzero group counts at each lambda step in
  the \code{\link{signpen}} object. The result is a two-column matrix with
  columns \code{Df} and \code{Lambda}. The \code{Df} column is the number of
  the groups that have nonzero within-group coefficients, the \code{Lambda}
  column is the the corresponding lambda.
}
\keyword{consistency}
\keyword{penalty}
\keyword{perceptron}
\keyword{sign}
